#20 move out i18n models and constants to specific folder and update their imports
