+++ src/app/i18n/i18n.constants.ts
@@ -1,4 +1,4 @@
-import { Labels, languages, labelsKeys } from '../models/i18n.models';
+import { Labels, languages, labelsKeys } from './i18n.models';
 import {
   COMMON_WARNING_MESSAGE,
   UNKNOWN_ERROR_MESSAGE,
